changing the location of files + minor edits
