+++ Assets/Academy-Platformer/Bootstrap/Interface/IBootstrap.cs
@@ -1,0 +1,9 @@
+﻿using Command;
+namespace Bootstrap
+{
+    public interface IBootstrap
+    {
+        void Add(ICommand command);
+        void Execute();
+    }
+}
