+++ Assets/Academy-Platformer/Command.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: e980394183aa444caeab446233377661
+timeCreated: 1690428202
