+++ Assets/Academy-Platformer/Command/Interface.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: c87a05c07b2c486cba8c17f87922a3e8
+timeCreated: 1690428215
