+++ Assets/Academy-Platformer/Command/Interface/ICommand.cs
@@ -1,0 +1,10 @@
+﻿using System;
+namespace Command
+{
+    public interface ICommand
+    {
+        event Action OnCommandExecuteNotify; 
+        void Execute();
+        void Undo();
+    }
+}
