--- Assets/Bootstrap/Interface/IBootstrap.cs
@@ -1,8 +1,0 @@
-﻿namespace Bootstrap.Interface
-{
-    public interface IBootstrap
-    {
-        public bool Add(ICommand command);
-        public bool Start();
-    }
-}
