--- Assets/Bootstrap/Interface/ICommand.cs
@@ -1,11 +1,0 @@
-﻿using System;
-
-namespace Bootstrap.Interface
-{
-    public interface ICommand
-    {
-        public event Action CommandExecuteNotify; 
-        public void Execute();
-        public void Undo();
-    }
-}
